requirements.txt
google-generativeai==0.8.3
coverage==7.6.0
click==8.1.7
astunparse==1.6.3
pytest==8.2.2
pytest-cov==5.0.0
flask